pyproject.toml
[tool.poetry]
name = "trading-fastapi"
version = "0.1.0"
description = ""
authors = ["novel <18.novel.18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
celery = "^5.3.4"
redis = "^5.0.1"
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"
kafka-python = "^2.0.2"
motor = "^3.3.1"
flake8 = "^6.1.0"
mypy = "^1.6.1"
pylint = "^3.0.2"
isort = "^5.12.0"
typing-extensions = "^4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
